+++ bcipy/helpers/demo/demo_visualization.py
@@ -21,7 +21,7 @@ from bcipy.helpers.acquisition import analysis_channels
 from bcipy.helpers.load import (load_experimental_data, load_json_parameters,
                                 load_raw_data)
 from bcipy.helpers.triggers import TriggerType, trigger_decoder
-from bcipy.helpers.visualization import visualize_gaze
+from bcipy.helpers.visualization import visualize_erp
 from bcipy.signal.process import get_default_transform
 
 if __name__ == '__main__':
@@ -91,7 +91,7 @@ if __name__ == '__main__':
 
     save_path = None if not args.save else path
 
-    figure_handles = visualize_gaze(
+    figure_handles = visualize_erp(
         raw_data,
         channel_map,
         trigger_timing,